backend/requirements.txt
Flask
python-dotenv
langchain-openai
langchain-community
langchain-core
langchain-text-splitters
langchain-chroma
unstructured
unstructured[md]
